Enum list_contacts::Error

#[non_exhaustive]
pub(crate) enum Error {

    AccountSuspendedException(AccountSuspendedException),
    AlreadyExistsException(AlreadyExistsException),
    BadRequestException(BadRequestException),
    ConcurrentModificationException(ConcurrentModificationException),
    ConflictException(ConflictException),
    InvalidNextTokenException(InvalidNextTokenException),
    LimitExceededException(LimitExceededException),
    MailFromDomainNotVerifiedException(MailFromDomainNotVerifiedException),
    MessageRejected(MessageRejected),
    NotFoundException(NotFoundException),
    SendingPausedException(SendingPausedException),
    TooManyRequestsException(TooManyRequestsException),
    Unhandled(Box<dyn Error + Send + Sync + 'static, Global>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccountSuspendedException(AccountSuspendedException)

The message can't be sent because the account's ability to send email has been permanently restricted.

Tuple Fields of AccountSuspendedException

0: AccountSuspendedException

AlreadyExistsException(AlreadyExistsException)

The resource specified in your request already exists.

Tuple Fields of AlreadyExistsException

0: AlreadyExistsException

BadRequestException(BadRequestException)

The input you provided is invalid.

Tuple Fields of BadRequestException

0: BadRequestException

ConcurrentModificationException(ConcurrentModificationException)

The resource is being modified by another operation or thread.

Tuple Fields of ConcurrentModificationException

0: ConcurrentModificationException

ConflictException(ConflictException)

If there is already an ongoing account details update under review.

Tuple Fields of ConflictException

0: ConflictException

InvalidNextTokenException(InvalidNextTokenException)

The specified request includes an invalid or expired token.

Tuple Fields of InvalidNextTokenException

0: InvalidNextTokenException

LimitExceededException(LimitExceededException)

There are too many instances of the specified resource type.

Tuple Fields of LimitExceededException

0: LimitExceededException

MailFromDomainNotVerifiedException(MailFromDomainNotVerifiedException)

The message can't be sent because the sending domain isn't verified.

Tuple Fields of MailFromDomainNotVerifiedException

0: MailFromDomainNotVerifiedException

MessageRejected(MessageRejected)

The message can't be sent because it contains invalid content.

Tuple Fields of MessageRejected

0: MessageRejected

NotFoundException(NotFoundException)

The resource you attempted to access doesn't exist.

Tuple Fields of NotFoundException

0: NotFoundException

SendingPausedException(SendingPausedException)

The message can't be sent because the account's ability to send email is currently paused.

Tuple Fields of SendingPausedException

0: SendingPausedException

TooManyRequestsException(TooManyRequestsException)

Too many requests have been made to the operation.

Tuple Fields of TooManyRequestsException

0: TooManyRequestsException

Unhandled(Box<dyn Error + Send + Sync + 'static, Global>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static, Global>